fix: 이벤트 생성버튼 터치 시 이벤트 생성 페이지로 이동, CreateEvent로 네이밍 변경, CreateEventView 분리
